Move segment and base logic to UrlBuilder from Router.
